requirements-ee.txt
ansible-core>=2.15,<2.16
ansible-lint>=6.20,<7.0
yamllint>=1.35,<2.0
jmespath>=1.0
netaddr>=1.3
passlib>=1.7
pyyaml>=6.0
pydantic>=2.0
jinja2>=3.1
pytest>=7.0
